# docs/specs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Stoneboyz CRM API
  version: 0.1.0
  description: |
    Source of truth for all API contracts.
    Edit this file before implementing any endpoint.
    packages/api-client is generated from this file; do not hand-edit it.
servers:
  - url: http://localhost:3001/api/v1
    description: Local development
  - url: https://api.stoneboyz.app/v1
    description: Production placeholder
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: better-auth.session_token
  parameters:
    CustomerId:
      name: customerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    QuoteId:
      name: quoteId
      in: path
      required: true
      description: UUID of the quote
      schema:
        type: string
        format: uuid
...
    AreaId:
      name: areaId
      in: path
      required: true
      description: UUID of the quote area
      schema:
        type: string
        format: uuid
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Request validation failed
        details:
          type: object
...
    PortalQuote:
      type: object
      required: [id, quoteNumber, title, status, customerId, customerName, discountCents, taxCents, subtotalCents, taxRateBps, shareToken, areas, lineItems]
      properties:
        id:
          type: string
          format: uuid
        quoteNumber:
          type: string
        title:
          type: string
        status:
          type: string
        customerId:
          type: string
          format: uuid
        customerName:
          type: string
        validUntil:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        termsAndConditions:
          type: string
          nullable: true
        discountCents:
          type: integer
        taxCents:
          type: integer
        subtotalCents:
          type: integer
        taxRateBps:
          type: integer
        shareToken:
          type: string
          format: uuid
        areas:
          type: array
          items:
            type: object
            required: [id, name, sortOrder]
            properties:
              id:
                type: string
                format: uuid
              name:
                type: string
              sortOrder:
                type: integer
        lineItems:
          type: array
          items:
            type: object
            required: [id, stoneType, qty, qtyUnit, unitPriceCents, laborPriceCents, sortOrder]
            properties:
              id:
                type: string
                format: uuid
              stoneType:
                type: string
              qty:
                type: number
              qtyUnit:
                type: string
              unitPriceCents:
                type: integer
              laborPriceCents:
                type: integer
              lengthIn:
                type: number
                nullable: true
              widthIn:
                type: number
                nullable: true
              thicknessCm:
                type: number
                nullable: true
              edgeProfile:
                type: string
                nullable: true
              notes:
                type: string
                nullable: true
              sortOrder:
                type: integer
              quoteAreaId:
                type: string
                format: uuid
                nullable: true
...
    ActorRequest:
      type: object
      properties:
...
    QuoteStatus:
      type: string
      enum: [draft, sent, accepted, rejected]
    QuoteLineItem:
      type: object
      required:
        - id
        - quoteId
        - quoteAreaId
        - sortOrder
        - stoneType
        - qty
        - qtyUnit
        - unitPriceCents
        - laborPriceCents
        - lineTotalCents
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        quoteId:
          type: string
          format: uuid
        quoteAreaId:
          type: string
          format: uuid
          nullable: true
        slabId:
          type: string
          format: uuid
          nullable: true
        sortOrder:
          type: integer
          default: 0
        stoneType:
          type: string
          example: Marble Calacatta
        lengthIn:
          type: number
          nullable: true
        widthIn:
          type: number
          nullable: true
        thicknessCm:
          type: number
          nullable: true
          example: 2
        edgeProfile:
          type: string
          nullable: true
          example: bullnose
        qty:
          type: number
          format: decimal
        qtyUnit:
          type: string
          example: sqm
        unitPriceCents:
          type: integer
          minimum: 0
        laborPriceCents:
          type: integer
          minimum: 0
          default: 0
        lineTotalCents:
          type: integer
          description: Computed by application — floor(qty * (unitPriceCents + laborPriceCents)). Not stored in DB.
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QuoteArea:
      type: object
      required:
        - id
        - quoteId
        - sortOrder
        - name
        - subtotalCents
        - measurementTotals
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        quoteId:
          type: string
          format: uuid
        sortOrder:
          type: integer
          default: 0
        name:
          type: string
        material:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        edgeProfile:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        subtotalCents:
          type: integer
        measurementTotals:
          $ref: '#/components/schemas/QuoteMeasurementAreaTotals'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QuoteMeasurementAreaTotals:
      type: object
      required: [pieceCount, countertopSqFt, finishedEdgeLinFt, splashSqFt, sinkCutoutCount, faucetHoleCount]
      properties:
        pieceCount:
          type: integer
        countertopSqFt:
          type: number
        finishedEdgeLinFt:
          type: number
        splashSqFt:
          type: number
        sinkCutoutCount:
          type: integer
        faucetHoleCount:
          type: integer
...
    Quote:
      type: object
      required:
        - id
        - customerId
        - quoteNumber
        - title
        - status
        - discountCents
        - taxRateBps
        - subtotalCents
        - totalCents
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
          nullable: true
        priceListId:
          type: string
          format: uuid
          nullable: true
        quoteNumber:
          type: string
          example: Q-2026-001
        title:
          type: string
        status:
          $ref: '#/components/schemas/QuoteStatus'
        validUntil:
          type: string
          format: date
          nullable: true
        subtotalCents:
          type: integer
          description: Computed by application — sum of all lineTotalCents. Not stored in DB.
        discountCents:
          type: integer
          minimum: 0
          default: 0
        taxRateBps:
          type: integer
          minimum: 0
          default: 0
          description: Tax rate in basis points (e.g. 1500 = 15%)
        totalCents:
          type: integer
          description: Computed by application — floor((subtotalCents - discountCents) * (1 + taxRateBps / 10000)). Not stored in DB.
        notes:
          type: string
          nullable: true
        termsAndConditions:
          type: string
          nullable: true
        sentAt:
          type: string
          format: date-time
          nullable: true
        acceptedAt:
          type: string
          format: date-time
          nullable: true
        rejectedAt:
          type: string
          format: date-time
          nullable: true
        archivedAt:
          type: string
          format: date-time
          nullable: true
        archivedByUserId:
          type: string
          format: uuid
          nullable: true
        shareToken:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QuoteWithLineItems:
      allOf:
        - $ref: '#/components/schemas/Quote'
        - type: object
          required: [lineItems, areas]
          properties:
            lineItems:
              type: array
              items:
                $ref: '#/components/schemas/QuoteLineItem'
            areas:
              type: array
              items:
                $ref: '#/components/schemas/QuoteArea'
...
    CreateQuoteAreaRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        sortOrder:
          type: integer
          default: 0
        material:
          type: string
        color:
          type: string
        edgeProfile:
          type: string
        notes:
          type: string
    UpdateQuoteAreaRequest:
      type: object
      description: At least one updatable field is required.
      minProperties: 2
      properties:
        name:
          type: string
          minLength: 1
        sortOrder:
          type: integer
        material:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        edgeProfile:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
...
    UpdateQuoteRequest:
      type: object
      description: At least one updatable field is required. Only allowed when quote is in draft status.
      minProperties: 2
      properties:
        title:
          type: string
          minLength: 1
        projectId:
          type: string
          format: uuid
          nullable: true
        priceListId:
          type: string
          format: uuid
          nullable: true
        validUntil:
          type: string
          format: date
          nullable: true
        discountCents:
          type: integer
          minimum: 0
        taxRateBps:
          type: integer
          minimum: 0
        notes:
          type: string
          nullable: true
        termsAndConditions:
          type: string
          nullable: true
...
security:
  - bearerAuth: []
paths:
...
  /customers/{customerId}/quotes/{quoteId}:
    get:
      operationId: getCustomerQuote
      summary: Get quote detail with embedded line items
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
      responses:
        '200':
          description: Quote returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteWithLineItems'
        '404':
          description: Quote or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      operationId: updateCustomerQuote
      summary: Update quote header fields (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuoteRequest'
      responses:
        '200':
          description: Quote updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quote'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quote or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /customers/{customerId}/quotes/{quoteId}/areas:
    get:
      operationId: listQuoteAreas
      summary: List all areas for a quote
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
      responses:
        '200':
          description: Areas returned.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuoteArea'
        '404':
          description: Quote or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      operationId: createQuoteArea
      summary: Create a quote area (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuoteAreaRequest'
      responses:
        '201':
          description: Area created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteArea'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Quote or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /customers/{customerId}/quotes/{quoteId}/areas/{areaId}:
    patch:
      operationId: updateQuoteArea
      summary: Update a quote area (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateQuoteAreaRequest'
      responses:
        '200':
          description: Area updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteArea'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      operationId: removeQuoteArea
      summary: Remove a quote area (draft only)
      parameters:
        - $ref: '#/components/parameters/CustomerId'
        - $ref: '#/components/parameters/QuoteId'
        - $ref: '#/components/parameters/AreaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActorRequest'
      responses:
        '200':
          description: Area removed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteArea'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Area, quote, or customer not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Area has line items or quote is not in draft status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /portal/quotes/{token}:
    get:
      operationId: getPortalQuote
      summary: Get quote by share token (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Quote for customer portal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PortalQuote'
        '404':
          description: Quote not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...